pyproject.toml
[tool.poetry]
name = "genshinapi"
version = "0.1.0"
description = "Unofficial Korean Genshin/Hoyolab API written in FastAPI"
authors = ["Leo Kim <me@leok.kr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
fastapi = "^0.68.1"
slowapi = "^0.1.4"
genshinstats = "^1.4.8"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
